Create realtime-specific ProductWithoutPrice card
